refactor storage policy
